Move Zeppelin support to okapi-api
